+++ packages/brick_graphql/lib/src/graphql_provider.dart
@@ -37,6 +37,9 @@ class GraphqlProvider extends Provider<GraphqlModel> {
     return Request(
       operation: Operation(document: defaultOperation.document),
       variables: query?.providerArgs['variables'] ?? variables ?? queryToVariables<_Model>(query),
+      context: query?.providerArgs['context'] != null
+          ? Context.fromMap(Map<Type, ContextEntry>.from(query?.providerArgs['context']))
+          : Context(),
     );
   }